disallow multiple invocations of start/stop for a given job and process
This allows `monit restart` to behave in a similar manner to
`monit stop && monit start`, specifically that start will wait for stop
to complete before attempting to start the process. This is because
`monit restart` forks `monit start` immediately after the pidfile
becomes invalid, instead of waiting for the result of the `monit stop`
command.

This uses a lock on a deterministic filepath:
`/var/vcap/sys/run/bpm/<job>/<process>.lock`. It has been implemented
using the underlying `SYS_FLOCK` syscall on unix (and therefore
currently does not work on other operating systems).

[finishes #148711965]

Signed-off-by: Rob Dimsdale-Zucker <rdimsdale@pivotal.io>