Add `capabilities` process configuration
This change introduces a new `capabilities` configuration that allows
users to configure an allowed set of linux capabilities for their
processes. Currently the only capability that we allow to be set is the
`CAP_NET_BIND_SERVICE` capability which allows processes to bind to
privileged ports. Capabilities should be added to this array of
capabilities without the preceding `CAP_`.

This change also adds `netcat-openbsd` to the docker image for the
integration tests.

[finishes #151259947]

Signed-off-by: James Myers <jfmyers9@gmail.com>