Add `ephemeral_disk` configuration
This change introduces a new `ephemeral_disk` boolean option. When set
to true, BPM will mount `/var/vcap/data/<job-name>` inside the process's
container.

If the key is not provided, or its value is false, the directory
will still exist inside the container as a side-effect of mounting the
temporary directory at `/var/vcap/data/<job-name>/tmp`, but it will have
no contents and will not be writable.

[finishes #151532030]

Signed-off-by: Rob Dimsdale-Zucker <rdimsdale@pivotal.io>