move logs from .{out,err}.log to .std{out,err}.log
This helps keep consistency with other BOSH logs (such as pre-start
logs) which use this convention. We do not believe anyone is depending
directly on the log file name. Any syslog forwarders and log rotators
should only be looking for the *.log suffix.

The commands used for this change were:

  grep -rl ".out.log" src/ docs/ | xargs sed -i 's/\.out\.log/.stdout.log/g'
  grep -rl ".err.log" src/ docs/ | xargs sed -i 's/\.err\.log/.stderr.log/g'