scripts: selectively use sudo to run docker
If the running user is a member of the docker group then sudo is not
required in order to write to the docker socket. This script now checks
for group membership to determine whether or not to run docker as root.

This also prevents the ~/.docker directory from being owned by root. You
may need to change the ownership of this directory back to your user
after this change (although the tests should still run without doing
that).

The command to add your user to the docker group is:

    sudo gpasswd -a $USER docker

Signed-off-by: Christopher Brown <cbrown@pivotal.io>