scripts: check docker permissions without groups
I originally wanted this fix because the privilege escalation is not
required on MacOS but ended up tunnel-visioning on the implementation
for the Linux workstation I was working on. This caused two issues:

  1) It encouraged users to add the docker group to their user which has
     a number of security caveats[1] as reported by @jfmyers9.

  2) The fix doesn't even work on MacOS.

This new implementation checks if a user can use sudo-less docker
commands by actually trying to run one rather than trying to infer it
from the environment. This addresses the two points above.

[1]: https://docs.docker.com/engine/security/security/#docker-daemon-attack-surface