update file copyright headers
The existing headers should have been changed when this was moved into
the Cloud Foundry Foundation (CFF). This change reallocates the
copyright from Pivotal to the CFF.

This change updates the licence checking test suite to be less picky
about the exact copyright header. We should start using 2018 in the new
files and the previous test would not allow this.

The command used to update the majority of these files was:

  grep -rl "Pivotal Software" src | xargs sed -i 's/Pivotal Software/CloudFoundry.org Foundation/g'