bpm: vendor packages
This changes uses dep to vendor all of our dependencies into our source
tree. The bpm directory is now in a state where we could move it into a
new repository if necessary. The submodules are still present for now
but will be removed in a future commit. They are no longer used.

* Teach scripts/sync-package-specs how to reference vendored code and
  regenerate the packaging specs.

* Vendor tools which we explicitly rely on (ginkgo, counterfeiter) and
  install them from the local copy rather than running go get at build
  time.

* Our licence checker should not check vendored code.