Include stopped processes in list command output
This change modifies the list command to include processes that are
defined in a job's `bpm.yml` that are not currently started. They will
show up in the `stopped` state. If the process is not running due to
failure, they will continue to be displayed in the `failed` state.

Processes will be shown in the order determined by the OS when fetching
the list of known jobs from `/var/vcap/jobs`. By default on linux this
will be in alphabetical order. Within a job, processes will be shown in
the order that they are defined in the `bpm.yml`.

When an invalid `bpm.yml` is present, `bpm list` will warn the user. It
will also warn the user when a runc process exists that is not defined
in any job's `bpm.yml`.

[finishes #152366823]