Add version command and global version flag
This change adds a `bpm version` command and a global `--version` flag
that can be used to display the version of bpm. The `src/version` file
controls the version displayed by this command when created deployed in
a BOSH release. The final release build will bump the value of this file
to the version that is being created, create a final release, and then
add a `+dev` to the currently released version.

[finishes #155400037]