Change group owner of `/etcd/profile.d/bpm.sh`
This change updates the group owner of `/etc/profile.d/bpm.sh` to
`vcap`. A recent `umask` change in the 3541.x stemcells, made the
`bpm.sh` script inaccessible to the `bosh ssh` user. This change will
allow non root users that are in the `vcap` group to add `bpm` to the
PATH when launching a login shell.

[finishes #156140254]