runc/lifecycle: use error to express missing processes
We modelled the renamed StatProcess function on the os.Stat function
from the standard library. This means that we can safely say that if the
error is nil then there will always be a process returned. The
IsNotExist function was added to check for the missing process error case.

This check was then added to the shell command so that the user does not
see an internal error message with internal container names.

* Re-ordered and re-grouped some imports so that bpm imports are at the
  bottom.

* Renamed job to process where necessary.

* Use the state of a process instead of its PID to see if it has failed.

* Do not try to start a job if there is an error trying to fetch
  existing process state.

[finishes #156173439]

Signed-off-by: Christopher Brown <cbrown@pivotal.io>