Add `unsafe.privileged` to process configuration
This change allows a user to request a privileged container using the
`unsafe.privileged` configuration option. A privileged container will:

  - launch the process as the root user
  - remove restrictions on syscalls through seccomp
  - enable a default set of extended privileges
  - remove masked and read only paths
  - remove restrictions on new privileges

The default set of capabilities is similar to other containerization
platforms.

This change also refactors the adapter to use a new `specbuilder`
package. This new package follows a popular style of using option
functions to modify an object in place while constructing it.

[finishes #156173174]