specbuilder: do not limit swap if unsupported
Some machines (notably some of the default GCP machine images) do not
support limiting swap usage out of the box. This can be irritating and
prevents clean integration test runs. This change now checks to see if
the host system supports swap before trying to set it.

The system information is injected in from the top level rather than
being encapsuated inside the specbuilder package to aid testing and to
avoid coupling the specbuilder package to only being able to build RunC
specs for the machine it is currently running on.

[fixes #156893141]