config: error if job uses reserved volume path
We only write this to the log since this will normally be run by monit
and standard out and error will not be visible to the user.

The adapter has logic removed from it and instead only worries about
what is actually possible at runtime. It should allow configurations
which are practically possible without worrying about what BPM
standards.

[finishes #157259659]