add run command
The bpm run command executes short-lived processes in the foreground.
This is useful for running BOSH errands and other one-off scripts. The
configuration is identical to long-running processes.

The exit status of the process is mimicked by BPM. A new exitstatus
package has been added in order to shove an exit status and an error
through an error shaped hole. This is useful at the top level when
working with our CLI library.

I found it very difficult to add integration tests for this feature in
the existing suite. The BDD style that Ginkgo brings is nice for unit
tests with few dependencies but makes integration tests very difficult
to read as the flow through the test ends up jumping all over the file
and sometimes takes non-obvious steps to push setup into different
levels of BeforeEach/JustBeforeEach to avoid restructuring test files
when adding new tests.

In an attempt to improve things I have started a new experimental
integration test suite called integration2. It currently only contains
tests for the run command. The testing package from the standard library
is used for an imperative flow while a set of sandbox functions reduces
the setup boilerplate.

This package is still an experiment and I'm looking for feedback on it.