Add support for unrestricted volumes
This change allows users to specify a list of volumes in the unsafe
configuration block. Volumes within this list are not subject to the
restriction that they must be within `/var/vcap/data` or
`/var/vcap/store`.

This change enables jobs with complex filesystem dependencies to be
converted to use bpm.

[#158122381]