config: remove restriction on allowed capabilities
We don't want people reaching for the privileged option when they don't
need to be. We could provide a validation for the capabilities to help
protect against typos. I don't know if we want the maintenance burden
yet to keep up with newly added capabilities. The error from RunC is
good and descriptive for now:

  starting container process caused "unknown capability \"CAP_NOT_REAL\""