lifecycle: decrease stop timeouts
We've seen these timeouts being reached when a job which ignores signals
is asked to stop on a heavily loaded machine. It can sometimes take over
5 seconds for a process to dissapear from the process table even when
being sent SIGKILL.

There isn't much we can do if the machine is so heavily loaded it can't
kill processes. We have a hard 30 second limit enforced by monit which
we want to adhere to.

[finishes #158202290]