Allow regular files to be specified in additional volumes
This change updates `bpm` to allow regular files to specified in the
additional volumes configuration. Previously, this would fail due to the
fact that `bpm` attempts to `mkdir -p` all volumes specified. Now when
an additional volume is an existing file, `bpm` will only attempt to
`chown` the file to the vcap user and group.

This change also refactors the adapter's `CreateJobPrerequisites` to be
more readable.

[finishes #157259699](https://www.pivotaltracker.com/story/show/157259699)