cgroup: improve cgroup mounting sophistication
Not all kernels will have all of the cgroups enabled so instead of
hard-coding a list of cgroups we should query the kernel to see which
groups are actually available.

Additionally, if we ourselves are already running in a cgroup then we
aren't able to mount different cgroups in different hierarchies to the
parent[1]. Therefore we must read our own cgroup groups and match those
with our own mounts.

[1]: https://lwn.net/Articles/606699/