integration: parse mount output rather than substring-ing
We saw test failures in some environments due to the string "rootfs" in
their mountpoint which unfortunately contains "ro". This led the tests
to believe that read-only option was set.

This change uses the newly exposed mount file parser to parse the output
which let's us write more precise assertions.