jobid: use semi human-readable container IDs
Even though we don't expose these as part of our public interface it can
be helpful for operators using container metrics collectors if our
container IDs are somewhat human readable.

The size of this changeset is disappointing. It would be nice if we
didn't depend on these functions directly in every test, and instead
there were helpers for creating container names in the integration
tests.