client: do not write pidfile for "bpm run"
Writing a pidfile is not necessary for "bpm run" as the job stays in the
foreground. This also meant that we never cleaned these pidfiles up and
so they would (harmlessly, but untidily) leak onto the system.

The `bpm pid` command will still work as it queries runc directly rather
than reading this file.

[fixes #163704234]