move tmpfs workaround from pre-start to bpm
This ensures that things continue to work if the machine is restarted,
the mount is lost, and pre-start isn't run. This change doesn't actually
fix any bugs in practice but I feel much better about the consistency of
doing it this way.

The /var/vcap/data/bpm/tmpworkaround path is only actually used on the
Trusty stemcell because it's missing the memfd_create syscall. If you're
using Xenial then it doesn't even matter if this is mounted or not.

Originally RunC used the path /tmp to create an anonymous file
descriptor to clone the init binary. We patched it to use this new
path because on BOSH-lite because /tmp was not mounted in such a way
that we could use the O_TMPFILE option. We mounted a tmpfs on the new
path to ensure that it would always work in all environments.

However, BOSH-lite instances cannot be restarted so pre-start will
always be run and so the mount will always be there. In real
environments we use ext4 for the /tmp mountpoint which supports the
O_TMPFILE option. The only thing we need to ensure is that the
underlying directory is still there. Since only the mount is lost on a
reboot and no on-disk state changes this will work.