Use `os.RemoveAll` instead of `os.Remove` to clean up pidfile
This change uses `RemoveAll` to forcibly delete the pidfile when
cleaning up a failed bpm process.

After a BOSH vm is rebooted, the contents of `/var/vcap/sys/run` are not
present. At the moment, bpm will fail to clean up the pidfile associated
with the failed process. This change fixes an error that occurs when
starting up a failed container after a reboot.