runc/adapter: use functional options for mounts
This makes it much easier to see which mount options are used beyond the
baseline options. The sea of string mount options did not make it clear
which options were different.

The BeMount and HaveMount matchers have also been added to help with
comparing mounts in tests. The order of the mount options does not
matter but our tests were fragile when the order changed.