bosh: require explicit internal/external paths
While working on injecting a custom init binary into BPM processes I ran
into a surprising pattern in our integration tests. The custom BOSH root
that we use to allow our tests to be concurrent was being used inside
the test BPM jobs too. e.g. instead of /var/vcap it would use something
akin to /tmp/bpmtest-{sandbox-id} both in and outside the test job.
Since we're already using mount namespaces we can avoid our tests from
needing to know they're running in a test-harness.

To this end, this change adds a new bosh.Path struct which can hold an
abstract path before being explicitly resolved to either an internal or
external path. This forces the developer to make a decision based on
where the path will be used and using this pattern across the codebase
has caused the problem described in the above paragraph to disappear
without much effort.

The bosh.Path struct is a monad for convenience. Paths can be appended
to without resolving the internal or external representation. This lets
them be passed throughout the program without needing to make a decision
on their usage until the last possible call-site.