flock: extract flock implementation
We're going to be using these locks in more places soon so now seems
like a good time to extract them into a slightly more re-usable form.

This change also fixes a subtle bug with the old implementation which
could cause a race condition which would invalidate the guarantees
provided by the lock. The bug is shown with three processes (A, B, and
C) and the followings steps:

   A opens file
   A locks file
   B opens file
   A unlinks file
   A unlocks file
   B locks file (B holds a lock on the deleted file)
   C opens file (a new file is created)
   C locks file (two processes hold locks on the same path!)

This can be avoided by not deleting the file and instead just issuing
the appropriate unlock flag to the flock(2) syscall. This does mean that
the user cannot tell if a lock is in-use or not by just looking at the
filesystem but in this case correctness trumps everything.