hostlock: allow placing locks on bpm objects
This higher-level interface is in preparation of allowing locks to be
placed on volumes as well a jobs. It also has the advantage of hiding
the implementation details of the locking from most of the codebase.
This let's us only expose the hostlock interface in the future which
allows swapping to POSIX locks or adding support for Windows locks etc.