runc/adapter: volumes can be marked as shared
If a volume is marked as shared then any updates to mounts inside it
will be reflected in all mount namespaces rather than just reflecting
the state of the mounts at job start time.

In order to avoid race conditions when creating and mounting the volume
host-level locking wraps all sharing operations.

The syntax for enabling shared volumes is:

  processes:
  - name : ...
    ...
    additional_volumes:
    - path: /var/vcap/data/shared
      shared: true