runc/adapter: inject a reaping init process
We've seen a common bug where users are somehow leaking zombie
processes. Our workaround for this up until now has been to recommend
that users use /bin/bash to start their process as it is able to reap
processes. Unfortunately bash doesn't forward signals to child processes
and so BPM is never able to cleanly stop these jobs.

In order to remove this source of errors, this commit adds a small init
process (https://github.com/krallin/tini) which is able to reap
processes and forward signals to the child process.

I've manually tested that this works but it could do with integration
test coverage too. I'm not yet sure how to test this with our current
test harness but that's the job for the next commit.

closes #121