integration: de-flake a shell test
It's possible for the output from the /bin/echo $TERM can take longer
than the standard Eventually(...) timeout on a heavily loaded machine.
We can't elegantly wait an indefinite amount of time for some output to
appear the same way that we can wait for a process to exit so instead we
combine both commands into one and wait for the process to exit.