Relax container does not exist regular expression
This change relaxes the regular expression for detecting whether or not
a container exists when shelling out to runc. In a recent release they
added more output to the command that caused this regular expression to
fail.

This change also updates the pid, shell, and trace commands to not panic
when an actual error is returned that is not of type `IsNotExist`.