Fix volume validation for `additional_volumes` (#138)
This change fixes the `pathIsIn` function to properly fail if a path is
not nested within one of the provided paths. This change is potentially
a breaking change due to the fact that this bug has been around for a
long period of time.

**Note**: This change also relaxes the validation for volumes to ensure that they
are nested within `/var/vcap`. This is an unfortunate side effect of
this bug as it will be a breaking change for a good number of users that
have been relying on the broken behavior.