delete broken runc state after restart
Sometimes the RunC state directory can be empty after a restart which
causes RunC to error when trying to start the same job. If the job fails
to start then we cleanup that directory so that when monit attempts to
restart the job it should succeed.

There are some open questions...

* This may have debug-ability issues by cleaning up the failed job?

* Maybe this means we should switch this directory to be on tmpfs?

Fixes #130