🐞 CI: Stop "No such file or directory" errors
The `stop_test.go` unit tests were unable to write their log files, triggering an
error in the `stderr.log` file.

On the surface the problem seemed odd: `stdout.log`, `stderr.log`, and
`foo.log` were being written in the same directory, so why was there a
problem?

The answer is that `stdout.log` and `stderr.log` were being written
_outside_ the container, which had a completely different filesystem
view than `foo.log` which was being written from _within_ the container.

The fix was copy the pattern from the `start_test.go` unit test.

For those curious, here are the paths as seen from inside & outside the
container:

| Path outside the container | /bpmtmp/stop-testGUID/sys/log/GUID/foo.log |
|----------------------------|--------------------------------------------|
| Path inside the container  | /var/vcap/sys/log/GUID/foo.log             |

Fixes:
```
/bin/bash: line 6: /bpmtmp/stop-test1655879386/sys/log/5b4af9df-8003-4a66-92ed-0ccec3adee0d/foo.log: No such file or directory
```