Use the docker-cpi image while testing
The bosh/integration does not have the `start-bosh` capabilities that
the bosh/main-bosh-docker image had, so it cannot run the tests.

As the bosh/main-bosh-docker image has an outdated golang version, we
cannot use it. This commit moves to the bosh/docker-cpi image that
could be a good fit for the testing tasks.