Use bosh golang-release for golang
Most bosh-related pipelines already use bosh/golang-release to make the
right version of golang available. This makes the BPM docker image
consistent with the rest of the ecosystem.

Additionally, there was a create-release task that used golang as its
image. However, it does not appear to be used anywhere, and thus is
deleted.