Increase memory limit for OOM test
The bump to runc 1.1.15 caused this test to fail consistently.  There is
an existing issue about k8s tests also failing, with the suggestion that
the memory limit be increased:
https://github.com/opencontainers/runc/issues/4427

In addition, a similar commit happened ~6 years ago:
https://github.com/cloudfoundry/bpm-release/commit/e4df8fb09a4aa4cc24934995a9723907ccd1bd48

We may be able to revert this change once more is understood about the
runc issue.