Randomize errand job name during tests to avoid cgroup conflicts
Tests in the integration2 package use the same `errand` job name when
executing `bpm run`. When these tests are run in quick succession, it
causes a race between cgroups from the previous test (with the same
name) being cleaned up after `bpm run` is complete and the currently
executing test. When `runc` tries to create a new container, it finds
a cgroup with the same name (errand) already exists and has processes
inside it, so it errors with a `runc run failed: container's cgroup is
not empty: 2 process(es) found` message.

This change adds a random string to the errand job's name, so cgroups do
not conflict with each other.

Signed-off-by: Nishad Mathur <nishad.mathur@broadcom.com>