Increase open file limit in spec
After upgrading to runc 1.3, 10 open files is too small to even launch BPM in
the "too many open files" test. As a result, we have increased it to 20 to
allow launch, and the corresponding bash script to write > 20 files.

It also appears that 1.3 reverts the changes that caused us to modify this test
in 6503cb7031964e53078bcd94fb725c20cfc9b92b, so we have restored the previous
assertions.

Signed-off-by: Julian Hjortshoj <julian.hjortshoj@broadcom.com>